Updating local

+++ Modules/Module/Http/Controllers/ModuleController.php
@@ -142,7 +142,7 @@ class ModuleController extends Controller
         $permission->name = $companyId.'.'.(config('config-variables.menu_item_permission_identifier')).'.'.$module->id;
         $permission->save();
 
-        $role = Role::find(1);
+        $role = Role::find(session('currentrole'));
         $role->givePermissionTo($permission);
 
         flash()->success(config('config-variables.flash_messages.dataSaved'));
@@ -205,7 +205,7 @@ class ModuleController extends Controller
         $type = 'success';
         $companyId = Landlord::getTenants()['company']->id;
 
-        Permission::where('name', $companyId.'.'.$moduleId)->delete();
+        Permission::where('name', $companyId.'.'.(config('config-variables.menu_item_permission_identifier')).'.'.$moduleId)->delete();
 
         if (!MenuItem::where('id', $moduleId)->delete()) {
             $message = config('config-variables.flash_messages.dataNotDeleted');
